+++ src/vkw/framebuffer.cpp
@@ -7,7 +7,7 @@ framebuffer::framebuffer(VkRenderPass renderPass, std::span<const VkImageView> v
     VkFramebufferCreateInfo framebuf_info{};
     framebuf_info.sType = VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO;
     framebuf_info.renderPass = renderPass;
-    framebuf_info.attachmentCount = views.size();
+    framebuf_info.attachmentCount = static_cast<uint32_t>(views.size());
     framebuf_info.pAttachments = views.data();
     framebuf_info.width = extent.width;
     framebuf_info.height = extent.height;
